Fix int/EMACS_INT in eval.c and indent.c, correct previous changes.
 indent.c (Fcurrent_indentation, indented_beyond_p)
 (compute_motion): Use EMACS_INT for buffer position variables.
 lisp.h (indented_beyond_p): Adjust prototype.
 buffer.c (overlay_strings): Return EMACS_INT.
 buffer.h (overlay_strings): Adjust prototype.
 region-cache.c (pp_cache): Adjust format to arguments.
 eval.c <specpdl_size, lisp_eval_depth>: Declare EMACS_INT.
 (call_debugger): Use EMACS_INT for specpdl_size related variables.
 (verror): Use EMACS_INT for size of allocated buffer.

+++ src/region-cache.c
@@ -788,9 +788,9 @@ pp_cache (struct region_cache *c)
   EMACS_INT end_u = c->buffer_end - c->end_unchanged;
 
   fprintf (stderr,
-           "basis: %d..%d    modified: %d..%d\n",
-           c->buffer_beg, c->buffer_end,
-           beg_u, end_u);
+           "basis: %ld..%ld    modified: %ld..%ld\n",
+           (long)c->buffer_beg, (long)c->buffer_end,
+           (long)beg_u, (long)end_u);
 
   for (i = 0; i < c->cache_len; i++)
     {
@@ -804,7 +804,7 @@ pp_cache (struct region_cache *c)
              : (pos == end_u) ? '-'
              : ' '),
             stderr);
-      fprintf (stderr, "%d : %d\n", pos, BOUNDARY_VALUE (c, i));
+      fprintf (stderr, "%ld : %d\n", (long)pos, BOUNDARY_VALUE (c, i));
     }
 }
